+++ src/main/java/com/brand/adapaxels/content/paxels/base/PaxelItem.java
@@ -2,78 +2,81 @@ package com.brand.adapaxels.content.paxels.base;
 
 import com.brand.adapaxels.tags.APBlockTags;
 import com.google.common.collect.BiMap;
+import net.fabricmc.fabric.mixin.content.registry.AxeItemAccessor;
 import net.fabricmc.fabric.mixin.content.registry.ShovelItemAccessor;
-import net.minecraft.advancement.criterion.Criteria;
-import net.minecraft.block.*;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.*;
-import net.minecraft.server.network.ServerPlayerEntity;
-import net.minecraft.sound.SoundCategory;
-import net.minecraft.sound.SoundEvents;
-import net.minecraft.util.ActionResult;
-import net.minecraft.util.Hand;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Direction;
-import net.minecraft.world.World;
-import net.minecraft.world.event.GameEvent;
+import net.minecraft.advancements.CriteriaTriggers;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.*;
+import net.minecraft.world.item.context.UseOnContext;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.CampfireBlock;
+import net.minecraft.world.level.block.RotatedPillarBlock;
+import net.minecraft.world.level.block.WeatheringCopper;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.gameevent.GameEvent;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Map;
 import java.util.Optional;
 
 @SuppressWarnings({"rawtypes"})
-public class PaxelItem extends MiningToolItem {
-    private final String materialName;
-    private static final Map<Block, BlockState> PATH_STATES;
+public class PaxelItem extends DiggerItem {
+    public final String materialName;
+    public static final Map<Block, BlockState> FLATTENABLES;
+    public static final Map<Block, Block> STRIPPABLES;
 
-    public PaxelItem(ToolMaterial material, Item.Settings settings, String materialName) {
-        super(material, APBlockTags.PAXEL_MINEABLE, settings);
+    public PaxelItem(Tier tier, Item.Properties properties, String materialName) {
+        super(tier, APBlockTags.MINEABLE_WITH_PAXEL, properties);
         this.materialName = materialName;
     }
 
-    public String getMaterialName() {
-        return materialName;
+    public InteractionResult useOn(UseOnContext context) {
+        Level level = context.getLevel();
+        BlockPos blockPos = context.getClickedPos();
+        Player player = context.getPlayer();
+
+        if (playerHasShieldUseIntent(context)) {
+            return InteractionResult.PASS;
+        }
+
+        if (stripBlock(context, level, blockPos, player)) {
+            return InteractionResult.sidedSuccess(level.isClientSide);
+        }
+
+        if (context.getClickedFace() == Direction.DOWN) {
+            return InteractionResult.PASS;
+        }
+
+        if (flattenOrExtinguishBlock(context, level, blockPos, player)) {
+            return InteractionResult.sidedSuccess(level.isClientSide);
+        }
+
+        return InteractionResult.PASS;
     }
 
-    public ActionResult useOnBlock(ItemUsageContext context) {
-        World world = context.getWorld();
-        BlockPos blockPos = context.getBlockPos();
-        PlayerEntity playerEntity = context.getPlayer();
-
-        if (shouldCancelStripAttempt(context)) {
-            return ActionResult.PASS;
-        }
-
-        if (stripBlock(context, world, blockPos, playerEntity)) {
-            return ActionResult.success(world.isClient);
-        }
-
-        if (context.getSide() == Direction.DOWN) {
-            return ActionResult.PASS;
-        }
-
-        if (flattenOrExtinguishBlock(context, world, blockPos, playerEntity)) {
-            return ActionResult.success(world.isClient);
-        }
-
-        return ActionResult.PASS;
-    }
-
-    private boolean stripBlock(ItemUsageContext context, World world, BlockPos blockPos, PlayerEntity playerEntity) {
-        Optional<BlockState> optional = this.tryStrip(world, blockPos, playerEntity, world.getBlockState(blockPos));
+    private boolean stripBlock(UseOnContext context, Level level, BlockPos blockPos, Player player) {
+        Optional<BlockState> optional = this.evaluateNewBlockState(level, blockPos, player, level.getBlockState(blockPos));
         if (optional.isPresent()) {
-            ItemStack itemStack = context.getStack();
-            if (playerEntity instanceof ServerPlayerEntity) {
-                Criteria.ITEM_USED_ON_BLOCK.trigger((ServerPlayerEntity) playerEntity, blockPos, itemStack);
+            ItemStack itemInHand = context.getItemInHand();
+            if (player instanceof ServerPlayer) {
+                CriteriaTriggers.ITEM_USED_ON_BLOCK.trigger((ServerPlayer) player, blockPos, itemInHand);
             }
 
             BlockState newState = optional.get();
-            world.setBlockState(blockPos, newState, 11);
-            world.emitGameEvent(GameEvent.BLOCK_CHANGE, blockPos, GameEvent.Emitter.of(playerEntity, newState));
+            level.setBlock(blockPos, newState, 11);
+            level.gameEvent(GameEvent.BLOCK_CHANGE, blockPos, GameEvent.Context.of(player, newState));
 
-            if (playerEntity != null) {
-                itemStack.damage(1, playerEntity, LivingEntity.getSlotForHand(context.getHand()));
+            if (player != null) {
+                itemInHand.hurtAndBreak(1, player, LivingEntity.getSlotForHand(context.getHand()));
             }
 
             return true;
@@ -81,58 +84,58 @@ public class PaxelItem extends MiningToolItem {
         return false;
     }
 
-    private boolean flattenOrExtinguishBlock(ItemUsageContext context, World world, BlockPos blockPos, PlayerEntity playerEntity) {
-        BlockState blockState = world.getBlockState(blockPos);
-        BlockState blockState2 = PATH_STATES.get(blockState.getBlock());
+    private boolean flattenOrExtinguishBlock(UseOnContext context, Level level, BlockPos blockPos, Player player) {
+        BlockState blockState = level.getBlockState(blockPos);
+        BlockState blockState2 = FLATTENABLES.get(blockState.getBlock());
 
-        if (blockState2 != null && world.getBlockState(blockPos.up()).isAir()) {
-            world.playSound(playerEntity, blockPos, SoundEvents.ITEM_SHOVEL_FLATTEN, SoundCategory.BLOCKS, 1.0F, 1.0F);
-            updateBlockState(world, blockPos, blockState2, playerEntity, context);
+        if (blockState2 != null && level.getBlockState(blockPos.above()).isAir()) {
+            level.playSound(player, blockPos, SoundEvents.SHOVEL_FLATTEN, SoundSource.BLOCKS, 1.0F, 1.0F);
+            updateBlockState(level, blockPos, blockState2, player, context);
             return true;
         }
 
-        if (blockState.getBlock() instanceof CampfireBlock && blockState.get(CampfireBlock.LIT)) {
-            if (!world.isClient()) {
-                world.syncWorldEvent(null, 1009, blockPos, 0);
+        if (blockState.getBlock() instanceof CampfireBlock && blockState.getValue(CampfireBlock.LIT)) {
+            if (!level.isClientSide()) {
+                level.levelEvent(null, 1009, blockPos, 0);
             }
-            CampfireBlock.extinguish(context.getPlayer(), world, blockPos, blockState);
-            BlockState blockState3 = blockState.with(CampfireBlock.LIT, false);
-            updateBlockState(world, blockPos, blockState3, playerEntity, context);
+            CampfireBlock.dowse(context.getPlayer(), level, blockPos, blockState);
+            BlockState blockState3 = blockState.setValue(CampfireBlock.LIT, false);
+            updateBlockState(level, blockPos, blockState3, player, context);
             return true;
         }
 
         return false;
     }
 
-    private void updateBlockState(World world, BlockPos blockPos, BlockState newState, PlayerEntity playerEntity, ItemUsageContext context) {
-        world.setBlockState(blockPos, newState, 11);
-        world.emitGameEvent(GameEvent.BLOCK_CHANGE, blockPos, GameEvent.Emitter.of(playerEntity, newState));
-        if (playerEntity != null) {
-            context.getStack().damage(1, playerEntity, LivingEntity.getSlotForHand(context.getHand()));
+    private void updateBlockState(Level level, BlockPos blockPos, BlockState newState, Player player, UseOnContext context) {
+        level.setBlock(blockPos, newState, 11);
+        level.gameEvent(GameEvent.BLOCK_CHANGE, blockPos, GameEvent.Context.of(player, newState));
+        if (player != null) {
+            context.getItemInHand().hurtAndBreak(1, player, LivingEntity.getSlotForHand(context.getHand()));
         }
     }
 
-    private static boolean shouldCancelStripAttempt(ItemUsageContext context) {
-        PlayerEntity playerEntity = context.getPlayer();
-        return context.getHand().equals(Hand.MAIN_HAND) && playerEntity.getOffHandStack().isOf(Items.SHIELD) && !playerEntity.shouldCancelInteraction();
+    private static boolean playerHasShieldUseIntent(UseOnContext context) {
+        Player player = context.getPlayer();
+        return context.getHand().equals(InteractionHand.MAIN_HAND) && player.getOffhandItem().is(Items.SHIELD) && !player.isSecondaryUseActive();
     }
 
-    private Optional<BlockState> tryStrip(World world, BlockPos pos, @Nullable PlayerEntity player, BlockState state) {
-        Optional<BlockState> optional = this.getStrippedState(state);
+    private Optional<BlockState> evaluateNewBlockState(Level level, BlockPos blockPos, @Nullable Player player, BlockState blockState) {
+        Optional<BlockState> optional = this.getStripped(blockState);
         if (optional.isPresent()) {
-            world.playSound(player, pos, SoundEvents.ITEM_AXE_STRIP, SoundCategory.BLOCKS, 1.0F, 1.0F);
+            level.playSound(player, blockPos, SoundEvents.AXE_STRIP, SoundSource.BLOCKS, 1.0F, 1.0F);
             return optional;
         } else {
-            Optional<BlockState> optional2 = Oxidizable.getDecreasedOxidationState(state);
+            Optional<BlockState> optional2 = WeatheringCopper.getPrevious(blockState);
             if (optional2.isPresent()) {
-                world.playSound(player, pos, SoundEvents.ITEM_AXE_SCRAPE, SoundCategory.BLOCKS, 1.0F, 1.0F);
-                world.syncWorldEvent(player, 3005, pos, 0);
+                level.playSound(player, blockPos, SoundEvents.AXE_SCRAPE, SoundSource.BLOCKS, 1.0F, 1.0F);
+                level.levelEvent(player, 3005, blockPos, 0);
                 return optional2;
             } else {
-                Optional<BlockState> optional3 = Optional.ofNullable((Block) ((BiMap) HoneycombItem.WAXED_TO_UNWAXED_BLOCKS.get()).get(state.getBlock())).map((block) -> block.getStateWithProperties(state));
+                Optional<BlockState> optional3 = Optional.ofNullable((Block) ((BiMap) HoneycombItem.WAX_OFF_BY_BLOCK.get()).get(blockState.getBlock())).map((block) -> block.withPropertiesOf(blockState));
                 if (optional3.isPresent()) {
-                    world.playSound(player, pos, SoundEvents.ITEM_AXE_WAX_OFF, SoundCategory.BLOCKS, 1.0F, 1.0F);
-                    world.syncWorldEvent(player, 3004, pos, 0);
+                    level.playSound(player, blockPos, SoundEvents.AXE_WAX_OFF, SoundSource.BLOCKS, 1.0F, 1.0F);
+                    level.levelEvent(player, 3004, blockPos, 0);
                     return optional3;
                 } else {
                     return Optional.empty();
@@ -141,11 +144,12 @@ public class PaxelItem extends MiningToolItem {
         }
     }
 
-    private Optional<BlockState> getStrippedState(BlockState state) {
-        return Optional.ofNullable(AxeItem.STRIPPED_BLOCKS.get(state.getBlock())).map((block) -> block.getDefaultState().with(PillarBlock.AXIS, state.get(PillarBlock.AXIS)));
+    private Optional<BlockState> getStripped(BlockState blockState) {
+        return Optional.ofNullable(STRIPPABLES.get(blockState.getBlock())).map((block) -> block.defaultBlockState().setValue(RotatedPillarBlock.AXIS, blockState.getValue(RotatedPillarBlock.AXIS)));
     }
 
     static {
-        PATH_STATES = ShovelItemAccessor.getPathStates();
+        FLATTENABLES = ShovelItemAccessor.getPathStates();
+        STRIPPABLES = AxeItemAccessor.getStrippedBlocks();
     }
 }