+++ src/main/java/com/brand/adapaxels/tags/APBlockTags.java
@@ -1,18 +1,18 @@
 package com.brand.adapaxels.tags;
 
 import com.brand.adapaxels.AdaPaxels;
-import net.minecraft.block.Block;
-import net.minecraft.registry.RegistryKeys;
-import net.minecraft.registry.tag.TagKey;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.level.block.Block;
 
 public class APBlockTags {
-    public static final TagKey<Block> PAXEL_MINEABLE = register("mineable/paxel");
+    public static final TagKey<Block> MINEABLE_WITH_PAXEL = register("mineable/paxel");
 
     private APBlockTags() {
     }
 
     private static TagKey<Block> register(String id) {
-        return TagKey.of(RegistryKeys.BLOCK, AdaPaxels.id(id));
+        return TagKey.create(Registries.BLOCK, AdaPaxels.id(id));
     }
 }