+++ changelog.md
@@ -1,1 +1,3 @@
-### Updated to **1.21.1**
+* Now supports Adabranium again
+* Updated Additional Additions paxel properties to reflect the latest version of the mod
+* Migrated code to Mojang mappings